2025/tool-sandbox/infra-slack-bot/pyproject.toml
[project]
name = "infra-slack-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "shooonng",email = "5511068t@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
